requirements.txt
amqp==5.3.1
asgiref==3.8.1
billiard==4.2.1
celery==5.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
Django==5.1.3
django-celery-results==2.5.1
djangorestframework==3.15.2
iniconfig==2.0.0
kombu==5.4.2
packaging==24.2
pika==1.3.2
pluggy==1.5.0
prompt_toolkit==3.0.48
psycopg==3.2.3
pytest==8.3.4
pytest-django==4.9.0
python-dateutil==2.9.0.post0
redis==5.2.0
setuptools==75.4.0
six==1.16.0
sqlparse==0.5.1
tzdata==2024.2
vine==5.1.0
wcwidth==0.2.13
